+++ src/inspect_ai/_eval/score.py
@@ -13,6 +13,13 @@ from inspect_ai.log import (
 )
 from inspect_ai.model import ModelName
 from inspect_ai.scorer import Metric, Score, Scorer, Target
+from inspect_ai.scorer._metric import SampleScore
+from inspect_ai.scorer._reducer import (
+    ScoreReducer,
+    ScoreReducers,
+    create_reducers,
+    reducer_log_names,
+)
 from inspect_ai.scorer._scorer import unique_scorer_name
 from inspect_ai.solver import TaskState
 
@@ -21,14 +28,19 @@ from .task.results import eval_results
 from .task.util import task_run_dir
 
 
-def score(log: EvalLog, scorers: Scorer | list[Scorer]) -> EvalLog:
+def score(
+    log: EvalLog,
+    scorers: Scorer | list[Scorer],
+    epochs_reducer: ScoreReducers | None = None,
+) -> EvalLog:
     """Score an evaluation log.
 
     Args:
        log (EvalLog): Evaluation log.
        scorers (Scorer): List of Scorers to apply to log
-       metrics: (list[Metric]): Additional metrics to compute
-         (Scorer built-in metrics are always computed).
+       epochs_reducer (ScoreReducers | None):
+           Reducer function(s) for aggregating scores in each sample.
+           Defaults to previously used reducer(s).
 
     Returns:
        Log with scores yielded by scorer.
@@ -39,15 +51,23 @@ def score(log: EvalLog, scorers: Scorer | list[Scorer]) -> EvalLog:
     # resolve scorers into a list
     scorers = [scorers] if isinstance(scorers, Scorer) else scorers
 
-    return asyncio.run(score_async(log, scorers))
+    return asyncio.run(score_async(log, scorers, epochs_reducer))
 
 
-async def score_async(log: EvalLog, scorers: list[Scorer]) -> EvalLog:
+async def score_async(
+    log: EvalLog,
+    scorers: list[Scorer],
+    epochs_reducer: ScoreReducers | None = None,
+) -> EvalLog:
     """Score an evaluation log.
 
     Args:
        log (EvalLog): Evaluation log.
        scorers (list[Scorer]): Scorers to apply to log
+       epochs_reducer (ScoreReducers  | None):
+         Reducer function(s) for aggregating scores in each sample.
+         Defaults to previously used reducer(s).
+
 
     Returns:
        Log with scores yielded by scorer.
@@ -85,18 +105,25 @@ async def score_async(log: EvalLog, scorers: list[Scorer]) -> EvalLog:
         ]
 
         # do scoring
-        scores: list[dict[str, Score]] = await asyncio.gather(*tasks)
+        scores: list[dict[str, SampleScore]] = await asyncio.gather(*tasks)
 
         # write them back (gather ensures that they come back in the same order)
         for index, score in enumerate(scores):
-            log.samples[index].scores = score
+            log.samples[index].scores = cast(dict[str, Score], score)
 
         # collect metrics from EvalLog (they may overlap w/ the scorer metrics,
         # that will be taken care of in eval_results)
         log_metrics = metrics_from_log(log)
 
+        # override epochs_reducer if specified
+        epochs_reducer = create_reducers(epochs_reducer)
+        if epochs_reducer:
+            log.eval.config.epochs_reducer = reducer_log_names(epochs_reducer)
+        else:
+            epochs_reducer = reducers_from_log(log)
+
         # compute metrics
-        log.results = eval_results(scores, scorers, log_metrics)
+        log.results = eval_results(scores, epochs_reducer, scorers, log_metrics)
 
     return log
 
@@ -120,8 +147,23 @@ async def task_score(task: Task, log: EvalLog) -> EvalLog:
     # compute and log metrics
     display().print(f"Aggregating scores for task: {task_name}")
     if task.scorer and log.samples:
+        sample_scores = [
+            {
+                score_key: SampleScore(
+                    sample_id=sample.id,
+                    value=score.value,
+                    answer=score.answer,
+                    explanation=score.explanation,
+                )
+                for score_key, score in sample.scores.items()
+            }
+            for sample in log.samples
+            if sample.scores is not None
+        ]
+
         log.results = eval_results(
-            [sample.scores for sample in log.samples if sample.scores is not None],
+            sample_scores,
+            task.epochs_reducer,
             task.scorer,
             task.metrics,
         )
@@ -133,12 +175,18 @@ async def run_score_task(
     target: Target,
     scorers: list[Scorer],
     progress: Callable[..., None],
-) -> dict[str, Score]:
-    results: dict[str, Score] = {}
+) -> dict[str, SampleScore]:
+    results: dict[str, SampleScore] = {}
     for scorer in scorers:
         result = await scorer(state, target)
         scorer_name = unique_scorer_name(scorer, list(results.keys()))
-        results[scorer_name] = result
+
+        results[scorer_name] = SampleScore(
+            sample_id=state.sample_id,
+            value=result.value,
+            answer=result.answer,
+            explanation=result.explanation,
+        )
 
     progress()
     return results
@@ -158,3 +206,7 @@ def metrics_from_log(log: EvalLog) -> list[Metric]:
 
 def metric_from_log(metric: EvalMetric) -> Metric:
     return cast(Metric, registry_create("metric", metric.name, **metric.options))
+
+
+def reducers_from_log(log: EvalLog) -> list[ScoreReducer] | None:
+    return create_reducers(log.eval.config.epochs_reducer)